backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
groq
python-multipart
Pillow
ultralytics